Showing result values in report
In this commit, we augmented our report by adding a table to show
the result values for each tracker. The values come from key
value(s) for each metric, such as 'AUC' from 'success_plot'. The
table is also sorted so we can see the performance order of
trackers easily.