Support new metric: "eao"
The new metric "eao", which is abbreviation for expected average
overlap, comes from VOT benchmark. In this commit, we implemented
a new plot "EAO curve" which shows the expected average overlap
between ground truth and estimated region under different sequence
length. We also calculated "EAO measure", which is the average EAO
for a typical range of sequence lengths. The value "EAO measure"
is only for calculated for the entrire dataset, not for videos.